+++ lexforge16/src/main/java/dev/nolij/zume/lexforge16/mixin/CameraMixin.java
@@ -11,10 +11,7 @@ public class CameraMixin {
 	
 	@ModifyArg(method = "setup", at = @At(value = "INVOKE", target = "Lnet/minecraft/client/Camera;getMaxZoom(D)D"))
 	public double zume$setup$getMaxZoom(double original) {
-		if (Zume.shouldHook())
-			return Zume.transformThirdPersonDistance(original);
-		
-		return original;
+        return Zume.transformThirdPersonDistance(original);
 	}
 	
 }